+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # Change Log
 
-## 0.0.1 (January 28, 2018)
-- Initial release. Syntax highlighting.
+### 0.3.0 (April 20, 2019)
+- Support for RNode installed locally
+- Fix v0.9.3 - error detection, cost output
 
-## 0.1.0 (February 02, 2019)
+### 0.2.0 (February 11, 2019)
+- Docker image user settings (default `rchain/rnode:master`)
+
+### 0.1.0 (February 02, 2019)
 - Grammar update for Rholang Mercury release v0.8>.
-- **Rholang Language Server** with error highlighting
+- **Rholang Language Server** with error highlighting.
+- Snippets for basic Rholang terms.
+
+### 0.0.1 (January 28, 2018)
+- Initial release. Syntax highlighting.